feat(client): add TabSkeletons component for wallet profile tabs
Add TabSkeletons component to render skeleton loading states for tokens, NFTs, and history tabs in the wallet profile view. Improves perceived performance during data loading by providing visual placeholders for tab content.
