+++ apps/client/src/components/panels/wallet/ProfileTabs/types.ts
@@ -1,0 +1,69 @@
+export type TabType = "tokens" | "nfts" | "history";
+
+export interface NFT {
+  id: string;
+  name: string;
+  collection: string;
+  image: string;
+  floorPrice: number;
+  tokenId?: string;
+  description?: string;
+}
+
+export interface Transaction {
+  hash: string;
+  to: string;
+  from: string;
+  data: string;
+  value: string;
+  isL1Originated: boolean;
+  fee: string;
+  nonce: number;
+  gasLimit: string;
+  gasPrice: string;
+  gasPerPubdata: string;
+  maxFeePerGas: string;
+  maxPriorityFeePerGas: string;
+  blockNumber: number;
+  l1BatchNumber: number;
+  blockHash: string;
+  type: number;
+  transactionIndex: number;
+  receivedAt: string;
+  error: string | null;
+  revertReason: string | null;
+  status: "included" | "pending" | "failed";
+  commitTxHash: string | null;
+  executeTxHash: string | null;
+  proveTxHash: string | null;
+  isL1BatchSealed: boolean;
+  gasUsed: string;
+  contractAddress: string | null;
+}
+
+export interface TransactionsMeta {
+  totalItems: number;
+  itemCount: number;
+  itemsPerPage: number;
+  totalPages: number;
+  currentPage: number;
+}
+
+export interface TransactionsLinks {
+  first: string;
+  previous: string;
+  next: string;
+  last: string;
+}
+
+export interface TransactionsResponse {
+  items: Transaction[];
+  meta: TransactionsMeta;
+  links: TransactionsLinks;
+}
+
+export interface ProfileTabsProps {
+  isEmpty?: boolean;
+  loading?: boolean;
+  onNavigateToExplore?: (tab?: "creators" | "pools" | "nfts") => void;
+}
